+++ spec/unit/dispatch.spec.ts
@@ -1,4 +1,4 @@
-import bus = require("../../src/");
+import Framebus = require("../../src/");
 import { dispatch } from "../../src/lib/dispatch";
 
 describe("dispatch", () => {
@@ -6,7 +6,9 @@ describe("dispatch", () => {
     const subscriber = jest.fn();
     const origin = "https://example.com";
 
-    bus.target(origin).on("test event", subscriber);
+    Framebus.target({
+      origin,
+    }).on("test event", subscriber);
 
     dispatch(origin, "test event", { data: "data" });
 
@@ -18,7 +20,9 @@ describe("dispatch", () => {
     const subscriber = jest.fn();
     const origin = "https://example.com";
 
-    bus.target(origin).on("test event", subscriber);
+    Framebus.target({
+      origin,
+    }).on("test event", subscriber);
 
     dispatch(origin, "different event", { data: "data" });
 
@@ -29,7 +33,9 @@ describe("dispatch", () => {
     const subscriber = jest.fn();
     const origin = "https://example.com";
 
-    bus.target(origin).on("test event", subscriber);
+    Framebus.target({
+      origin,
+    }).on("test event", subscriber);
 
     dispatch("https://domain.com", "test event", { data: "data" });
 
@@ -41,7 +47,9 @@ describe("dispatch", () => {
     const reply = jest.fn();
     const origin = "https://example.com";
 
-    bus.target(origin).on("test event", subscriber);
+    Framebus.target({
+      origin,
+    }).on("test event", subscriber);
 
     dispatch(origin, "test event", { data: "data" }, reply);
 
@@ -54,7 +62,9 @@ describe("dispatch", () => {
     const reply = jest.fn();
     const origin = "https://example.com";
 
-    bus.target(origin).on("test event", subscriber);
+    Framebus.target({
+      origin,
+    }).on("test event", subscriber);
 
     // eslint-disable-next-line no-undefined
     dispatch(origin, "test event", undefined, reply);