+++ spec/unit/emit.spec.ts
@@ -1,9 +1,10 @@
-import { attach } from "../../src/lib/attach";
-import bus = require("../../src/");
+import Framebus = require("../../src/");
 
 describe("emit", () => {
+  let bus: Framebus;
+
   beforeEach(() => {
-    attach();
+    bus = new Framebus();
   });
 
   it("should return false if event is not a string", () => {
@@ -15,8 +16,12 @@ describe("emit", () => {
 
   it("should return false if origin is not a string", () => {
     const actual = bus
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      .target({ origin: "object" } as any)
+      .target({
+        origin: {
+          foo: "object",
+        },
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      } as any)
       .emit("event", { data: "data" });
 
     expect(actual).toBe(false);
@@ -24,14 +29,20 @@ describe("emit", () => {
 
   it("should return true if origin and event are strings", () => {
     const actual = bus
-      .target("https://example.com")
+      .target({
+        origin: "https://example.com",
+      })
       .emit("event", { data: "data" });
 
     expect(actual).toBe(true);
   });
 
   it("can pass a reply function without passing data", () => {
-    const actual = bus.target("https://example.com").emit("event", jest.fn());
+    const actual = bus
+      .target({
+        origin: "https://example.com",
+      })
+      .emit("event", jest.fn());
 
     expect(actual).toBe(true);
   });