+++ spec/unit/helpers.ts
@@ -1,6 +1,0 @@
-import { detach } from "../../src/lib/attach";
-
-afterEach(() => {
-  jest.restoreAllMocks();
-  detach();
-});