+++ spec/unit/on.spec.ts
@@ -1,13 +1,23 @@
 import { subscribers } from "../../src/lib/constants";
-import bus = require("../../src/");
+import Framebus = require("../../src/");
 
 describe("on", () => {
+  let bus: Framebus;
+
+  beforeEach(() => {
+    bus = new Framebus();
+  });
+
   it("should add subscriber to given event and origin", () => {
     const event = "event name";
     const origin = "https://example.com";
     const fn = jest.fn();
 
-    bus.target(origin).on(event, fn);
+    bus
+      .target({
+        origin,
+      })
+      .on(event, fn);
 
     expect(subscribers[origin][event]).toEqual(expect.arrayContaining([fn]));
   });