+++ src/index.ts
@@ -1,8 +1,6 @@
 import { attach } from "./lib/attach";
 import { Framebus } from "./framebus";
 
-const bus = new Framebus();
-
 attach();
 
-export = bus;
+export = Framebus;